requirements.txt
streamlit==1.35.0
langchain==0.2.0
langchain-groq==0.1.3
langchain-neo4j==0.0.1
neo4j==5.20.0
pandas==2.2.1
openpyxl==3.1.2
python-dotenv==1.0.1  # 新增：读取.env文件